Setup the server.js entry point
Renamed the old as backup for now as reference as needed. Planning to remove once working.
